Fix new line on encoding folded style
When using:

 >-
   someline
     more

The encoder would emit:

 >-
   someline

     more

The issue was introduced in the original translation from C to
Go. In the C code the MOVE macro would advance the start pointer
of the string. In the Go case "k = 0" points to the start of the
string where in the C case it would be the read pointer. Fix it.

Also address bound checks to not go beyond the string (I assume in
the C code there is an explicit NUL or at least some NUL).

Fixes: #804